Reimplement "All Assets" button on new client_assets UI
I found it annoying that the new UI did not include a way to deselect the asset type filter. I have reimplemented All Assets button that works the same way as before. If this is not acceptable, a check to see if the button is currently active switches the href to remove the type when it's clicked again works. 

<a href="?client_id=<?php echo $client_id; ?><?php if ($_GET['type'] != 'workstation') { echo '&type=workstation'; } ?>"  class="btn <?php if ($_GET['type'] == 'workstation') { echo 'btn-primary'; } else { echo 'btn-default'; } ?>"><i class="fa fa-fw fa-desktop mr-2"></i>Workstations<span class="right badge badge-light ml-2"><?php echo $workstation_count; ?></span></a>